requirements.txt
pandas==1.5.3
tabula_py==2.9.0
